\name{testfn}
\alias{testfn}

\title{Fn for leak test}

\description{

Calls C++ function makegkPointcpp

}

\usage{

testfn(grain = 1, ncores = 2, parameters = c(0.5, 20))

}

\arguments{
  \item{grain}{integer}
  \item{ncores}{integer}
  \item{parameters}{numeric vector length 2)
}

}

\value{

Meaningless sum of probabilities gk

}

\examples{

testfn()

}
