\name{testfn}
\alias{testfn}

\title{Fn for leak test}

\description{

Calls C++ function makegkPointcpp with arbitrary (fake) data

}

\usage{

testfn(grain = 1, ncores = 2)

}

\arguments{
  \item{grain}{integer}
  \item{ncores}{integer}
}

}

\value{

Meaningless sum of probabilities gk

}

\examples{

testfn()

}
